requirements.txt
twscrape
pandas
loguru
python-telegram-bot
aiohttp
python-dotenv